re #3 re #4

+++ Storm.SvgMagic/SvgMagicOptions.cs
@@ -21,12 +21,12 @@ namespace Storm.SvgMagic
             bool.TryParse(queryStringParams.Get("force"), out force);
             options.Force = force;
 
-            int height;
-            int.TryParse(queryStringParams.Get("height"), out height);
+            float height;
+            float.TryParse(queryStringParams.Get("height"), out height);
             options.Height = height;
 
-            int width;
-            int.TryParse(queryStringParams.Get("width"), out width);
+            float width;
+            float.TryParse(queryStringParams.Get("width"), out width);
             options.Width = width;
 
             options.SetImageFormat(config);
@@ -51,8 +51,8 @@ namespace Storm.SvgMagic
         public string MimeType { get; set; }
         public string Extension { get; set; }
 
-        public int Height { get; set; }
-        public int Width { get; set; }
+        public float Height { get; set; }
+        public float Width { get; set; }
 
         public bool HasDimensions()
         {
